Started rebasing bot, that means:
Want to redo the way of receiving requests. Before I wanted to created
my own server to receive requests, but now I want to use existing tools.
Since now I'm creating new branch to try new version of bot that uses
looped way of receiving requests. It will change the project to the
root and it will work and be started in an absolutely different way.